Add DefaultRetries to test client too
